Get A Changing View Without Continual Reloading.
This whole frames of Turbo is actually kind of bonkers and mind melding
versus a true SPA. Get something that removes and updates the DOM as
expected when a transition from route to another occurs.

In this case it is a simple click from one watch reference to its'
detail.