A More 'Rails Like' Master/Detail View
Instead of having per watch turbo-frame targets, move to a single frame
that is rehydrated as websockets deliver each HTML payload. Selecting a
different watch will wipe the internals of this frame with the newly
selected watch info.

Overall this feels way more natural and allows for a seemingly easy
understanding of what information belongs in #show and #index.