Ignore New Environment File.
Moving most of the infrastructure (where possible) to Docker. Due to
this, there is a need to explicitely set the Database URL locally. This
will be meshed with the database.yaml config and therefore 'should'
allow the application to work as before. These environment items are
specific to the user, keep them out of source control.